Update CMakeLists.txt
Remove configure for retdec - it is requiring OpenSSL and other dependencies which makes no sense for just building the plugin.

The whereami library is already included as source code and the CMakeLists.txt chain includes it via dependencies.  So no action is needed for that.  It seems to build just fine without the configure option without having to fetch OpenSSL.