Fix code viewer leaking/improve terminate handling
The custom viewer is the parent and IDA does not close the tree for you.  It merely provides events for top level windows closing, and as for all the child widgets/viewers and such, one must manually close them to free the resources.  It took me some time to figure out exactly how this model works.  But I believe some simple changes will further make this resilient.